python/src/authentication/requirements.txt
asgiref==3.6.0
astroid==2.15.0
attrs==22.2.0
coverage==7.2.2
dill==0.3.6
Django==4.1.7
django-cors-headers==3.14.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.1
inflection==0.5.1
isort==5.12.0
jedi==0.18.2
jsonschema==4.17.3
lazy-object-proxy==1.9.0
mccabe==0.7.0
mysqlclient==2.1.1
parso==0.8.3
platformdirs==3.1.1
PyJWT==2.6.0
pylint==2.17.0
pyrsistent==0.19.3
python-decouple==3.8
pytz==2022.7.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tomlkit==0.11.6
uritemplate==4.1.1
wrapt==1.15.0
